SimilarityMicroservice/requirements.txt
Django==5.2
djangorestframework==3.15.2
django_cors_headers==4.6.0
python-decouple==3.8
dj-database-url==2.3.0
sentence-transformers~=4.1.0